factored scan browser widget into separate file
